+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.2 (2014/06/18)
+
+Enhancements:
+
+* Add default :quote_char, :name_sep for SQL::Maker::Select, and SQL::Maker::Condition
+* Add auto_bind option to SQL::Maker::Condition
+* Inherit SQL::Maker::Condition parameters on creating sub conditions
+
 ## 0.0.1 (2014/06/18)
 
 First version